+++ src/bencode.h
@@ -9,7 +9,7 @@
 #define BC_EOF			'e'
 
 
-int bc_decode_type(char c);
+uint8_t bc_decode_type(char c);
 size_t bc_read_lenstr(char *in, size_t *len);
 size_t bc_read_string(char *in, char **str);
 size_t bc_read_int(char *in, uint32_t *val);
